Add stats view and cloud function
Have to consider this approach before continuing. I was hoping that using cloud functions would provide it access to the contents of the firestore but it does not. Firestore contents would need to be retrieved with every cloud function call and as such, ignoring data usage, downloading the firestore contents once and performing analysis locally could be a better option. Another approach could be to maintain metrics in firestore which would be altered with CRUD operation. This would remove the need for any analysis to be done, and would mean only these metrics need to be retrieved rather than the entire attempts collection but would lead to added complexity for all firestore operations.